Refactor scope matching and model mutation planning
* Use more utilities in the tests
* Extract implementation code out of public API files
* Prepare scope matching for reuse
