+++ platforms/core-configuration/declarative-dsl-core/src/main/kotlin/org/gradle/internal/declarativedsl/dom/mutation/ModelMutation.kt
@@ -20,9 +20,6 @@ import org.gradle.declarative.dsl.schema.DataProperty
 import org.gradle.internal.declarativedsl.dom.DeclarativeDocument
 import org.gradle.internal.declarativedsl.dom.DeclarativeDocument.DocumentNode
 import org.gradle.internal.declarativedsl.dom.DeclarativeDocument.ValueNode
-import org.gradle.internal.declarativedsl.dom.DocumentResolution
-import org.gradle.internal.declarativedsl.dom.mutation.ModelMutationFailureReason.ScopeLocationNotMatched
-import org.gradle.internal.declarativedsl.dom.mutation.ModelMutationFailureReason.TargetPropertyNotFound
 import org.gradle.internal.declarativedsl.dom.resolution.DocumentResolutionContainer
 
 
@@ -62,7 +59,7 @@ sealed interface ModelMutation {
     ) : ModelPropertyMutation
 
     data class AddElement(
-        val newElement: DocumentNode.ElementNode,
+        val newElement: NewElementNodeProvider,
         val ifPresentBehavior: IfPresentBehavior
     ) : ModelMutation
 
@@ -78,10 +75,10 @@ sealed interface ModelMutation {
 }
 
 
-sealed interface NewDocumentNodeProvider {
-    data class Constant(val documentNode: DocumentNode) : NewDocumentNodeProvider
-    fun interface ArgumentBased : NewDocumentNodeProvider {
-        fun produceDocumentNode(argumentContainer: MutationArgumentContainer): DocumentNode
+sealed interface NewElementNodeProvider {
+    data class Constant(val elementNode: DocumentNode.ElementNode) : NewElementNodeProvider
+    fun interface ArgumentBased : NewElementNodeProvider {
+        fun produceElementNode(argumentContainer: MutationArgumentContainer): DocumentNode.ElementNode
     }
 }
 
@@ -104,110 +101,6 @@ sealed interface ModelMutationFailureReason {
     data object TargetPropertyNotFound : ModelMutationFailureReason
 
     data object ScopeLocationNotMatched : ModelMutationFailureReason
-}
-
-
-internal
-class DefaultModelToDocumentMutationPlanner : ModelToDocumentMutationPlanner {
-    override fun planModelMutation(
-        document: DeclarativeDocument,
-        resolution: DocumentResolutionContainer,
-        mutationRequest: ModelMutationRequest,
-        mutationArguments: MutationArgumentContainer
-    ): ModelMutationPlan =
-        toDocumentMutations(document, resolution, mutationRequest, mutationArguments)
-
-    private
-    fun toDocumentMutations(
-        document: DeclarativeDocument,
-        resolution: DocumentResolutionContainer,
-        request: ModelMutationRequest,
-        mutationArguments: MutationArgumentContainer
-    ): ModelMutationPlan =
-        when (request.mutation) {
-            is ModelMutation.UnsetProperty ->
-                withMatchingProperties(request, document, resolution) {
-                    DocumentMutation.DocumentNodeTargetedMutation.RemoveNode(it)
-                }
-
-            is ModelMutation.SetPropertyValue ->
-                withMatchingProperties(request, document, resolution) {
-                    DocumentMutation.ValueTargetedMutation.ReplaceValue(
-                        it.value,
-                        request.mutation.newValue.value(mutationArguments)
-                    )
-                }
-
-            is ModelMutation.AddElement -> withMatchingScopes(request, document, resolution) { matchingScopes ->
-                DefaultModelMutationPlan(
-                    matchingScopes
-                        .filter { it.elements.isNotEmpty() } // TODO: this is the top level scope, will need special handling
-                        .map { scope -> scope.elements.last().elementNodes }
-                        .map { elementResolution ->
-                            DocumentMutation.DocumentNodeTargetedMutation.ElementNodeMutation.AddChildrenToEndOfBlock(
-                                elementResolution.first,
-                                listOf(request.mutation.newElement)
-                            )
-                        }.toList(),
-                    emptyList()
-                )
-            }
-        }
-
-    private
-    fun NewValueNodeProvider.value(argumentContainer: MutationArgumentContainer) = when (this) {
-        is NewValueNodeProvider.ArgumentBased -> produceValueNode(argumentContainer)
-        is NewValueNodeProvider.Constant -> valueNode
-    }
-
-    private
-    fun withMatchingScopes(
-        request: ModelMutationRequest,
-        document: DeclarativeDocument,
-        resolution: DocumentResolutionContainer,
-        mapScopesToMutationPlan: (Set<Scope>) -> ModelMutationPlan
-    ): ModelMutationPlan {
-        val matchingScopes = request.location.match(document, resolution)
-        return when {
-            matchingScopes.isEmpty() -> DefaultModelMutationPlan(emptyList(), listOf(UnsuccessfulModelMutation(request, listOf(ScopeLocationNotMatched))))
-            else -> mapScopesToMutationPlan(matchingScopes)
-        }
-    }
-
-    private
-    fun withMatchingProperties(
-        request: ModelMutationRequest,
-        document: DeclarativeDocument,
-        resolution: DocumentResolutionContainer,
-        mapPropertyToDocumentMutation: (DocumentNode.PropertyNode) -> DocumentMutation
-    ): ModelMutationPlan {
-        return withMatchingScopes(request, document, resolution) { matchingScopes ->
-            val candidatePropertyNodes = matchingScopes
-                .filter { it.elements.isNotEmpty() } // TODO: this is the top level scope, will need special handling
-                .map { scope -> scope.elements.last() }
-                .flatMap { lastScopeElement ->
-                    lastScopeElement.elementNodes.first.content.filterIsInstance<DocumentNode.PropertyNode>()
-                }
-                .toSet()
-
-
-            val targetProperty = (request.mutation as ModelMutation.ModelPropertyMutation).property
-            val matchedPropertyNodes = candidatePropertyNodes
-                .filter {
-                    val propertyResolution = resolution.data(it) as DocumentResolution.PropertyResolution.PropertyAssignmentResolved // TODO: handle when this cast fails
-                    propertyResolution.property === targetProperty // TODO: why does simple equals not work here?
-                }
-
-            when {
-                matchedPropertyNodes.isEmpty() -> DefaultModelMutationPlan(
-                    emptyList(),
-                    listOf(UnsuccessfulModelMutation(request, listOf(TargetPropertyNotFound)))
-                )
-
-                else -> DefaultModelMutationPlan(matchedPropertyNodes.map(mapPropertyToDocumentMutation), emptyList())
-            }
-        }
-    }
 }
