+++ platforms/core-configuration/declarative-dsl-core/src/main/kotlin/org/gradle/internal/declarativedsl/dom/resolution/DefaultDocumentResolutionContainer.kt
@@ -22,11 +22,13 @@ import org.gradle.declarative.dsl.schema.AnalysisSchema
 import org.gradle.declarative.dsl.schema.DataClass
 import org.gradle.declarative.dsl.schema.FunctionSemantics
 import org.gradle.declarative.dsl.schema.SchemaMemberFunction
+import org.gradle.internal.declarativedsl.analysis.DefaultOperationGenerationId
 import org.gradle.internal.declarativedsl.analysis.ErrorReason
 import org.gradle.internal.declarativedsl.analysis.ObjectOrigin
 import org.gradle.internal.declarativedsl.analysis.ResolutionError
 import org.gradle.internal.declarativedsl.analysis.ResolutionTrace
 import org.gradle.internal.declarativedsl.analysis.SchemaTypeRefContext
+import org.gradle.internal.declarativedsl.analysis.analyzeEverything
 import org.gradle.internal.declarativedsl.analysis.getDataType
 import org.gradle.internal.declarativedsl.analysis.tracingCodeResolver
 import org.gradle.internal.declarativedsl.dom.AmbiguousName
@@ -66,8 +68,8 @@ class DocumentWithResolution(
 fun documentWithResolution(
     schema: AnalysisSchema,
     languageTreeResult: LanguageTreeResult,
-    operationGenerationId: OperationGenerationId,
-    analysisStatementFilter: AnalysisStatementFilter,
+    operationGenerationId: OperationGenerationId = DefaultOperationGenerationId.finalEvaluation,
+    analysisStatementFilter: AnalysisStatementFilter = analyzeEverything,
     strictReceiverChecks: Boolean = true
 ): DocumentWithResolution {
     val document = languageTreeResult.toDocument()
